+++ samples/NETFX-OldCSProj/Exports.cs
@@ -7,7 +7,7 @@ namespace NETFX_OldCSProj
         /// <summary>
         /// The exported c function will be named "DoSomething"
         /// </summary>
-        [Export]
+        [DllExport]
         public static void DoSomething()
         {
             System.Console.WriteLine(nameof(DoSomething));
@@ -16,7 +16,7 @@ namespace NETFX_OldCSProj
         /// <summary>
         /// The exported c function will be named "PINVOKE_Rocks"
         /// </summary>
-        [Export("PINVOKE_Rocks")]
+        [DllExport("PINVOKE_Rocks")]
         public static void DoSomething2ElectricBogaloo()
         {
             System.Console.WriteLine(nameof(DoSomething2ElectricBogaloo));