+++ samples/NETFX_SDKCSProj/Exports.cs
@@ -16,7 +16,7 @@ namespace NETFX_SDKCSProj
         /// <summary>
         /// The exported c function will be named "PINVOKE_Rocks"
         /// </summary>
-        [Export("PINVOKE_Rocks")]
+        [DllExport("PINVOKE_Rocks")]
         public static void DoSomething2ElectricBogaloo()
         {
             System.Console.WriteLine(nameof(DoSomething2ElectricBogaloo));