+++ tests/NXPorts.Tests/TestFiles/Simple.cs
@@ -5,7 +5,7 @@ namespace NXPorts.Tests.TestFiles
 {
     public static class Simple
     {
-        [Export]
+        [DllExport]
         public static void DoSomething()
         {
             Console.WriteLine("Test");