+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 <!-- markdownlint-disable MD041 -->
+- [BREAKINGCHANGE] #7 Renamed `ExportAttribute` to `DllExportAttribute`